reference/requirements.txt
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0  # For running tests
pdfplumber>=0.10.0  # PDF parsing with table extraction
sqlparse>=0.4.0  # SQL validation and formatting
pdf2image>=1.16.0  # Convert PDF pages to images
Pillow>=10.0.0  # Image processing
aiohttp>=3.9.0  # Async HTTP requests
tqdm>=4.66.0  # Progress bars
